moved class back to sanitize.ts to resolve errors
