+++ MVP/src/Processors/urlProcessor.ts
@@ -1,4 +1,4 @@
-import { CleanURLSet } from '../Types/URLTypes';
+import { CleanURLSet } from '../Input/Sanitize';
 import { Repository } from '../Types/DataTypes';
 import { processGitHubUrl } from './githubProcessor';
 import { processNpmUrl } from './registryProcessor';
