+++ MVP/src/index.ts
@@ -8,11 +8,11 @@ import { BaseRepoQueryResponse, ReposFromQuery } from './Types/ResponseTypes';
 import { requestFromGQL } from './Requests/GitHub/gql';
 import { mapGQLResultToRepos } from './Processors/gqlProcessor';
 import { DEFAULT_URLFILEPATH } from './Input/Input';
-import * as Sanitizer from './Input/Sanitize';
 import { LogMessage } from './Utils/log';
 import { ErrorWrapper, ErrorWrapperForAsync, ErrorWrapperForReturns } from './Utils/errorHandling';
+import { ProvideURLsForQuerying } from './Input/Sanitize';
 
-const cleanUrls = Sanitizer.ProvideURLsForQuerying(DEFAULT_URLFILEPATH, true);
+const cleanUrls = ProvideURLsForQuerying(DEFAULT_URLFILEPATH, true);
 console.log(cleanUrls.github_URLs);
 console.log(cleanUrls.npm_URLs);
 
@@ -22,7 +22,7 @@ console.log(`🌟 Everything appears to be ${chalk.greenBright('Operational')}! 
 catchArgs();
 
 const runner = async () => {
-    const cleanUrls = Sanitizer.ProvideURLsForQuerying(DEFAULT_URLFILEPATH, true);
+    const cleanUrls = ProvideURLsForQuerying(DEFAULT_URLFILEPATH, true);
     const repos = await buildReposFromUrls<BaseRepoQueryResponse>(cleanUrls); //using mock urls for now
     const query = repoQueryBuilder(repos); //add an array of fields here... see Request/QueryBuilders/fields.ts for examples
     const result = await requestFromGQL<ReposFromQuery<BaseRepoQueryResponse>>(query); //result is the raw gql response... .data has your data, .errors has the errors
